+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 - Fix levelup items not being usable on Tamed Wolves (or other mobs that override right click behavior)
 - Added config values for nameplate handling
 - Some internal javadoc cleanup
+- Use CanRender event over mixin (now fixed)
+- Fix Neoforge not having mod menu config screen
 
 ## 1.1.0
 - Allow Entities or Dimensions to set a player scaling multiplier, or disable biome/structure/player bonuses entirely. See README for usage